# backend/src/api/openapi.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Poker-API
  description: Progrendszerek projekt backend apija (Poker oldal no money)
...
paths:
...
  /checkAuth:
    get:
      summary: Check if user is authenticated
      operationId: checkAuth
      x-eov-operation-handler: api/controllers/auth.controller
      tags:
        - Auth
      responses:
        '200':
          description: User is authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Unauthorized, user not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
# Users 
...
  /chats:
    get:
      summary: Get all chat channels the user has access to
      operationId: getChats
      x-eov-operation-handler: api/controllers/chat.controller
      tags:
        - Chat
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: The page number to retrieve, chat channels are sorted by last message createdAt
      responses:
        '200':
          description: Chat channels retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatsResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
        - cookieAdminAuth: []
  /chat:
    put:
      summary: Update chat channel
      operationId: updateChat
      x-eov-operation-handler: api/controllers/chat.controller
      tags:
        - Chat
      parameters:
        - in: query
          name: chatId
          schema:
            type: string
          required: true
          description: The id of the chat channel to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                options:
                  type: object
                  properties:
                    key:
                      type: string
                    whiteList:
                      type: array
                      items:
                        type: string
                    banList:
                      type: array
                      items:
                        type: string
              required:
                - options
      responses:
        '200':
          description: Chat channel updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '404':
          description: Chat channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - chatOwnerAuth: []
        - cookieAdminAuth: []
    delete:
      summary: Delete chat channel
      operationId: deleteChat
      x-eov-operation-handler: api/controllers/chat.controller
      tags:
        - Chat
      parameters:
        - in: query
          name: chatId
          schema:
            type: string
          required: true
          description: The id of the chat channel to delete
      responses:
        '200':
          description: Chat channel deleted successfully
        '404':
          description: Chat channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - chatOwnerAuth: []
        - cookieAdminAuth: []
    post:
      summary: Create a new chat channel
      operationId: createChat
      x-eov-operation-handler: api/controllers/chat.controller
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                options:
                  type: object
                  properties:
                    key:
                      type: string
                    whiteList:
                      type: array
                      items:
                        type: string
                    banList:
                      type: array
                      items:
                        type: string
              required:
                - options
      responses:
        '200':
          description: Chat channel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
  /chat/message:
    get:
      summary: Get all messages of chat channel
      operationId: getMessages
      x-eov-operation-handler: api/controllers/chat.controller
      tags:
        - Chat
      parameters:
        - in: query
          name: channelId
          schema:
            type: string
          required: true
          description: The id of the chat channel to retrieve messages of
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: The page number to retrieve, messages are sorted by createdAt
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - chatMemberAuth: []
        - cookieAdminAuth: []
  /chat/join:
    get:
      summary: Join chat channel
      operationId: joinChat
      x-eov-operation-handler: api/controllers/chat.controller
      tags:
        - Chat
      parameters:
        - in: query
          name: channelId
          schema:
            type: string
          required: true
          description: The id of the chat channel to join
        - in: query
          name: token
          schema:
            type: string
          required: true
          description: The magic token to join the chat channel with
      responses:
        '200':
          description: Chat channel joined successfully
        '404':
          description: Chat channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
  /chat/leave:
    get:
      summary: Leave chat channel
      operationId: leaveChat
      x-eov-operation-handler: api/controllers/chat.controller
      tags:
        - Chat
      parameters:
        - in: query
          name: channelId
          schema:
            type: string
          required: true
          description: The id of the chat channel to leave
      responses:
        '200':
          description: Chat channel left successfully
        '404':
          description: Chat channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - chatMemberAuth: []
components:
  schemas:
...
    LoginResponse:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        balance:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
        - _id
...
    ChatsResponse:
      type: array
      items:
        $ref: '#/components/schemas/ChatResponse'
    ChatResponse:
      type: object
      properties:
        _id:
          type: string
        ownerId:
          type: string
        whiteList:
          type: array
          items:
            type: string
        banList:
          type: array
          items:
            type: string
        key:
          type: string
    MessageResponse:
      type: object
      properties:
        _id:
          type: string
        channelId:
          type: string
        senderId:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
    MessagesResponse:
      type: array
      items:
        $ref: '#/components/schemas/MessageResponse'
    BaseError:
      type: object
      properties:
        name:
          type: string
        message:
          type: string
        status:
          type: number
        originatedFrom:
          type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    userSpecificCookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    chatOwnerAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    chatMemberAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    gameOwnerAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    gamePlayerAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    cookieAdminAuth:
      type: apiKey
      in: cookie
      name: connect.sid

# TODO: mindehol jo response
# TODO: error messagek ahol lehet
# TODO: example es description
# TODO: plural or not plural, also generic all endpoint or single endpoint?
# FIXME: change query to path param where needed